requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (async)
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0          # async Postgres driver
psycopg2-binary==2.9.9   # sync driver (Alembic offline mode)
aiosqlite==0.20.0        # SQLite driver for tests

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9  # OAuth2PasswordRequestForm needs this

# Validation / Settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-dotenv==1.0.1

# Testing
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0